fix: make arrow function return explicit for objects
Objects returned from an arrow function should include an explicit return so that we are clear when we are returning an object. Otherwise, it can be difficult to identify where/when we are returning a value from an arrow function.

When not returning an object, we can omit curly braces for brevity.